More cleanups and docs.
